+++ ios/Classes/lib/utils/BitmapUtils.swift
@@ -70,15 +70,57 @@ public class BitmapUtils : MediaUtils {
             do {
                 
                 let imageData = try Data(contentsOf: url)
+                
+                // This is necessary to protect the target executions
+                // from EXC_RESOURCE_RESOURCE_TYPE_MEMORY fatal exception
+                
+                if SwiftUtils.isRunningOnExtension() && imageData.count > 1048576 {
+                    throw PushNotificationError.exceptionMsg(msg:
+                    "Notification image '\( String(describing: bitmapUri))' exceeds 1Mb"
+                    )
+                }
+                
                 return UIImage(data: imageData)
                 
             } catch let error {
                 print("error \(error)")
             }
         }
-
+        
         return nil
     }
+    
+    public static func resizeImage(image: UIImage, targetSize: CGSize) -> UIImage? {
+        let size = image.size
+
+        do {
+            let widthRatio  = targetSize.width  / size.width
+            let heightRatio = targetSize.height / size.height
+
+            // Figure out what our orientation is, and use that to form the rectangle
+            var newSize: CGSize
+            if(widthRatio > heightRatio) {
+                newSize = CGSize(width: size.width * heightRatio, height: size.height * heightRatio)
+            } else {
+                newSize = CGSize(width: size.width * widthRatio,  height: size.height * widthRatio)
+            }
+
+            // This is the rect that we've calculated out and this is what is actually used below
+            let rect = CGRect(x: 0, y: 0, width: newSize.width, height: newSize.height)
+
+            // Actually do the resizing to the rect using the ImageContext stuff
+            UIGraphicsBeginImageContextWithOptions(newSize, false, 0.9)
+            image.draw(in: rect)
+            let newImage = UIGraphicsGetImageFromCurrentImageContext()
+            UIGraphicsEndImageContext()
+
+            return newImage
+        }
+        catch let error  {
+            print("error \(error)")
+            return nil
+        }
+   }
     
     public static func getBitmapFromFile(_ mediaPath:String) -> UIImage? {
         let mediaPath:String? = BitmapUtils.cleanMediaPath(mediaPath)