+++ lib/src/utils/date_utils.dart
@@ -1,6 +1,9 @@
+import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
+import 'package:flutter/material.dart';
 
-class DateUtils {
+extension DateUtil on DateUtils {
+
   static DateTime parseStringToDate(String date,
       {String format: 'yyyy-MM-dd HH:mm:ss'}) {
     if (date == null || date.isEmpty) return null;